# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/mentors/{mentor_id}": {
            "get": {
                "tags": [
                    "Mentor"
                ],
                "summary": "Получение ментора по ID.",
                "operationId": "get_mentor_by_id_api_mentors__mentor_id__get",
                "parameters": [
                    {
                        "name": "mentor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Mentor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ментор успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mentor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ментор не найден"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mentors": {
            "get": {
                "tags": [
                    "Mentor"
                ],
                "summary": "Поиск менторов по запросу.",
                "operationId": "get_mentors_api_mentors_get",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "title": "Query"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список менторов по запросу получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mentor"
                                    },
                                    "title": "Response 200 Get Mentors Api Mentors Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mentors/{mentor_id}/score": {
            "post": {
                "tags": [
                    "Mentor"
                ],
                "summary": "Изменение рейтинга ментора.",
                "operationId": "add_score_api_mentors__mentor_id__score_post",
                "parameters": [
                    {
                        "name": "mentor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Mentor Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_add_score_api_mentors__mentor_id__score_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Скор ментора успешно изменен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mentor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ментор не найден"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mentors/sign-up": {
            "post": {
                "tags": [
                    "Mentor"
                ],
                "summary": "Регистрация ментора.",
                "operationId": "sign_up_mentor_api_mentors_sign_up_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MentorRegister"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Ментор успешно зарегистрирован.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MentorRegisterStatus"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ментор уже был зарегистрирован"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mentors/sign-in": {
            "post": {
                "tags": [
                    "Mentor"
                ],
                "summary": "Вход ментора.",
                "operationId": "sign_in_mentor_api_mentors_sign_in_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MentorSignIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Ментор успешно залогинился.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MentorRegisterStatus"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Неверные учётные данные"
                    },
                    "404": {
                        "description": "Ментор не найден."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/sign-up": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Регистрация админа",
                "operationId": "register_admin_api_admin_sign_up_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Успешная регистрация админа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminRegisterStatus"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/admin/sign-in": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Регистрация админа",
                "operationId": "login_admin_api_admin_sign_in_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация админа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminRegisterStatus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Админ не найден"
                    },
                    "403": {
                        "description": "Неверные данные для входа"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminCreate": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email",
                        "description": "Емейл админа"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password",
                        "description": "Пароль админа"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "AdminCreate"
            },
            "AdminRegisterStatus": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "title": "Status",
                        "description": "Статус добавления true/false",
                        "examples": [
                            true,
                            false
                        ]
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Message добавления",
                        "default": "ok"
                    },
                    "token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token",
                        "description": "JWT token"
                    },
                    "id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id",
                        "description": "UUID сущности"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "AdminRegisterStatus"
            },
...
            "Body_add_score_api_mentors__mentor_id__score_post": {
                "properties": {
                    "score": {
                        "type": "integer",
                        "maximum": 10.0,
                        "minimum": 0.0,
                        "title": "Score"
                    }
                },
                "type": "object",
                "required": [
                    "score"
                ],
                "title": "Body_add_score_api_mentors__mentor_id__score_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Mentor": {
                "properties": {
                    "mentor_id": {
                        "type": "string",
                        "title": "Mentor Id",
                        "description": "ID ментора"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Имя ментора"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email",
                        "description": "Email ментора"
                    },
                    "telegram": {
                        "type": "string",
                        "title": "Telegram",
                        "description": "Телеграм ментора"
                    },
                    "expertise": {
                        "type": "string",
                        "title": "Expertise",
                        "description": "Опыт ментора"
                    },
                    "bio": {
                        "type": "string",
                        "title": "Bio",
                        "description": "О себе"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score",
                        "description": "Рейтинг ментора",
                        "default": 0
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "description": "Теги ментора"
                    }
                },
                "type": "object",
                "required": [
                    "mentor_id",
                    "name",
                    "email",
                    "telegram",
                    "expertise",
                    "bio",
                    "tags"
                ],
                "title": "Mentor"
            },
            "MentorRegister": {
                "properties": {
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Имя ментора"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email",
                        "description": "Email ментора"
                    },
                    "telegram": {
                        "type": "string",
                        "title": "Telegram",
                        "description": "Телеграм ментора"
                    },
                    "expertise": {
                        "type": "string",
                        "title": "Expertise",
                        "description": "Опыт ментора"
                    },
                    "bio": {
                        "type": "string",
                        "title": "Bio",
                        "description": "О себе"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score",
                        "description": "Рейтинг ментора",
                        "default": 0
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "description": "Теги ментора"
                    }
                },
                "type": "object",
                "required": [
                    "password",
                    "name",
                    "email",
                    "telegram",
                    "expertise",
                    "bio",
                    "tags"
                ],
                "title": "MentorRegister"
            },
            "MentorRegisterStatus": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "title": "Status",
                        "description": "Статус добавления true/false"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Message добавления",
                        "default": "ok"
                    },
                    "token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token",
                        "description": "JWT token"
                    },
                    "id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id",
                        "description": "UUID сущности"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "MentorRegisterStatus"
            },
            "MentorSignIn": {
                "properties": {
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "password",
                    "email"
                ],
                "title": "MentorSignIn"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "CustomHTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}